Switch to std::uint_least32_t for stored indices in MatrixIndexSet
Using `std::uint_least32_t` instead of `std::size_t` halves the required memory
and thus also improves performance.
Notice that `std::uint32_t` is optional and `std::uint_fast32_t` is intended to be
fast in terms of computations and may have 64 bits. Here we're interested in reducing
memory and thus bandwith. Hence `std::uint_least32_t`, which is the smallest sufficient
type is most appropriate.

+++ dune/istl/matrixindexset.hh
@@ -31,7 +31,7 @@ namespace Dune {
    */
   class MatrixIndexSet
   {
-    using Index = std::size_t;
+    using Index = std::uint_least32_t;
 
     // A vector that partly mimics a std::set by staying
     // sorted on insert() and having unique values.
